@@ -25,18 +25,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "13.3.11",
-    "@angular/animations": "13.4.0",
-    "@angular/cdk": "13.3.9",
-    "@angular/common": "13.4.0",
-    "@angular/compiler": "13.4.0",
-    "@angular/core": "13.4.0",
-    "@angular/forms": "13.4.0",
-    "@angular/material": "13.3.9",
-    "@angular/platform-browser": "13.4.0",
-    "@angular/platform-browser-dynamic": "13.4.0",
-    "@fortawesome/angular-fontawesome": "0.10.2",
-    "@fortawesome/fontawesome-common-types": "0.2.36",
+    "@angular-devkit/build-angular": "14.2.11",
+    "@angular/animations": "14.3.0",
+    "@angular/cdk": "14.2.7",
+    "@angular/common": "14.3.0",
+    "@angular/compiler": "14.3.0",
+    "@angular/core": "14.3.0",
+    "@angular/forms": "14.3.0",
+    "@angular/material": "14.2.7",
+    "@angular/platform-browser": "14.3.0",
+    "@angular/platform-browser-dynamic": "14.3.0",
+    "@fortawesome/angular-fontawesome": "0.11.1",
+    "@fortawesome/fontawesome-common-types": "6.1.0",
     "@fortawesome/fontawesome-svg-core": "6.1.0",
     "@fortawesome/free-regular-svg-icons": "6.1.0",
     "@fortawesome/free-solid-svg-icons": "6.1.0",
@@ -55,9 +55,9 @@
   },
   "devDependencies": {
     "@angular-eslint/eslint-plugin": "~14.4",
-    "@angular/cli": "13.3.11",
-    "@angular/compiler-cli": "13.4.0",
-    "@angular/language-service": "13.4.0",
+    "@angular/cli": "14.2.11",
+    "@angular/compiler-cli": "14.3.0",
+    "@angular/language-service": "14.3.0",
     "@types/node": "16.18.34",
     "@typescript-eslint/eslint-plugin": "~5.59",
     "@typescript-eslint/parser": "~5.59",
@@ -72,10 +72,10 @@
     "npm-run-all": "4.1.5",
     "postcss": "~8.4",
     "postcss-sass": "0.5.0",
-    "stylelint": "14.16.1",
-    "stylelint-config-prettier": "9.0.5",
-    "stylelint-config-sass-guidelines": "9.0.1",
-    "stylelint-config-standard": "26.0.0",
+    "stylelint": "~15.7",
+    "stylelint-config-sass-guidelines": "~10.0",
+    "stylelint-config-standard": "~33.0",
+    "stylelint-config-standard-scss": "^9.0.0",
     "typescript": "~4.6.4"
   },
   "lint-staged": {
